Explicitly set seed datatype for windows compatibility

Former-commit-id: 7bdc5701a5d9b0ca3f3c0c062af0a02a575a0eef

+++ eemeter/eemeter/models/hourly/settings.py
@@ -346,7 +346,7 @@ class BaseHourlySettings(BaseSettings):
     @pydantic.model_validator(mode="after")
     def _check_seed(self):
         if self.seed is None:
-            self._seed = np.random.randint(0, 2**32 - 1)
+            self._seed = np.random.randint(0, 2**32 - 1, dtype=np.int64)
         else:
             self._seed = self.seed
